+++ flutter/lib/src/web_platform_checker.dart
@@ -1,2 +1,5 @@
 /// always false for flutter web
 bool isIOS() => false;
+
+/// always false for flutter web
+bool isAndroid() => false;