+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Fix: missing toList for debug meta #192
 - Enhancement: NavigationObserver to record Breadcrumbs for navigation events #197
 - Fix: add an integration closing system
+- Feat: Support split-debug-info for Android #191
+- Fix: the event payload must never serialize null or empty fields
 
 ### Breaking changes